Change vec3 to template class

+++ aabb.cpp
@@ -36,7 +36,7 @@ bool AABB::Hit(const ray& r, double t_min, double t_max) const
 
 double AABB::SurfaceArea(void) const
 {
-	vec3 v = maxp-minp;
+	dvec3 v = maxp-minp;
 	return 2.0*(v[0]*v[1]+v[1]*v[2]+v[2]*v[0]);
 }
 
@@ -44,10 +44,10 @@ double AABB::SurfaceArea(void) const
 
 AABB SurroundingBox(const AABB& box0, const AABB& box1)
 {
-	vec3 min_p(std::min(box0.minp.x(), box1.minp.x()),
+	dvec3 min_p(std::min(box0.minp.x(), box1.minp.x()),
 	std::min(box0.minp.y(), box1.minp.y()),
 	std::min(box0.minp.z(), box1.minp.z()));
-	vec3 max_p(std::max(box0.maxp.x(), box1.maxp.x()),
+	dvec3 max_p(std::max(box0.maxp.x(), box1.maxp.x()),
 	std::max(box0.maxp.y(), box1.maxp.y()),
 	std::max(box0.maxp.z(), box1.maxp.z()));
 	return AABB(min_p, max_p);
