+++ material.cpp
@@ -4,7 +4,7 @@
 #include "object.h"
 #include "stb_image.h"
 
-const vec3 default_vt(0, 0, 0);
+const dvec3 default_vt(0, 0, 0);
 
 // the approximation of shlick assumes that n is greater than or equal to 1
 double shlick(double cos_theta, double n)
@@ -39,12 +39,12 @@ double newshlick(double cos_theta, double n, double k)
 }
 
 
-double Homogenious::Phase(const vec3& vi, double wli, const vec3& vo, double wlo) const
+double Homogenious::Phase(const dvec3& vi, double wli, const dvec3& vo, double wlo) const
 {
 	return 1.0/(4.0 * M_PI);
 }
 
-bool Homogenious::Sample_p(const vec3& vo, double wlo, vec3& vi, double& wli, double& phase, double& pdfval) const
+bool Homogenious::Sample_p(const dvec3& vo, double wlo, dvec3& vi, double& wli, double& phase, double& pdfval) const
 {
 	double phi = 2.0 * M_PI * drand48();
 	double cos_theta = 1.0 - 2.0 * drand48();
@@ -52,21 +52,21 @@ bool Homogenious::Sample_p(const vec3& vo, double wlo, vec3& vi, double& wli, do
 	double x = sin(phi) * cos_theta;
 	double y = sin(phi) * sin_theta;
 	double z = cos(phi);
-	vi = vec3(x, y, z);
+	vi = dvec3(x, y, z);
 	wli = wlo;
 	phase = 1.0/(4.0 * M_PI);
 	pdfval = 1.0/(4.0 * M_PI);
 	return true;
 }
 
-double HenyeyGreenstein::Phase(const vec3& vi, double wli, const vec3& vo, double wlo) const
+double HenyeyGreenstein::Phase(const dvec3& vi, double wli, const dvec3& vo, double wlo) const
 {
 	const double cos_theta = dot(vi, vo);
 	const double denom = 1.0+g*g+2.0*g*cos_theta;
 	return (1.0-g*g)/(denom*sqrt(denom)*(4.0 * M_PI));
 }
 
-bool HenyeyGreenstein::Sample_p(const vec3& vo, double wlo, vec3& vi, double& wli, double& phase, double& pdfval) const
+bool HenyeyGreenstein::Sample_p(const dvec3& vo, double wlo, dvec3& vi, double& wli, double& phase, double& pdfval) const
 {
 	double phi = 2.0 * M_PI * drand48();
 	double square = (1.0-g*g)/(1.0-g+2.0*g*drand48());
@@ -75,7 +75,7 @@ bool HenyeyGreenstein::Sample_p(const vec3& vo, double wlo, vec3& vi, double& wl
 	double x = sin(phi) * cos_theta;
 	double y = sin(phi) * sin_theta;
 	double z = cos(phi);
-	vi = vec3(x, y, z);
+	vi = dvec3(x, y, z);
 	wli = wlo;
 	phase = this->Phase(vi, wli, vo, wlo);
 	pdfval = phase;
@@ -85,42 +85,42 @@ bool HenyeyGreenstein::Sample_p(const vec3& vo, double wlo, vec3& vi, double& wl
 
 //std::vector<std::shared_ptr<Hittable> > Material::lights;
 
-vec3 random_in_unit_Sphere(void)
+dvec3 random_in_unit_Sphere(void)
 {
-	vec3 p;
+	dvec3 p;
 	do {
-		p = 2.0*vec3(drand48(), drand48(), drand48()) - vec3(1, 1, 1);
+		p = 2.0*dvec3(drand48(), drand48(), drand48()) - dvec3(1, 1, 1);
 	} while (p.squared_length() >= 1.0);
 
 	return p;
 }
 
-vec3 reflect2(vec3 v, vec3 normal)
+dvec3 reflect2(dvec3 v, dvec3 normal)
 {
-	return (-v + 2 * normal * dot(v, normal));
+	return (-v + 2.0 * normal * dot(v, normal));
 }
 
-vec3 refract(vec3 v, vec3 normal, double n) // this n is a relative refraction index
+dvec3 refract(dvec3 v, dvec3 normal, double n) // this n is a relative refraction index
 {
-	//vec3 normal = unit_vector(rec.normal);
-	vec3 v2 = normal * dot(v, normal);
-	vec3 v1 = v - v2;
-	vec3 v1_p = -(1.0/n) * v1.length() * unit_vector(v1);
-	vec3 v2_p = -sqrt(1-v1_p.length()*v1_p.length()) * normal;
+	//dvec3 normal = unit_vector(rec.normal);
+	dvec3 v2 = normal * dot(v, normal);
+	dvec3 v1 = v - v2;
+	dvec3 v1_p = -(1.0/n) * v1.length() * unit_vector(v1);
+	dvec3 v2_p = -sqrt(1-v1_p.length()*v1_p.length()) * normal;
 	return v1_p + v2_p;
 }
-//vec3 refract(vec3 v, vec3 normal, double n_in, double n_out)
+//dvec3 refract(dvec3 v, dvec3 normal, double n_in, double n_out)
 //{
-//	//vec3 normal = unit_vector(rec.normal);
-//	vec3 v2 = normal * dot(v, normal);
-//	vec3 v1 = v - v2;
-//	vec3 v1_p = (n_in / n_out) * v1.length() * unit_vector(v1);
-//	vec3 v2_p = -sqrt(1-v1_p.length()*v1_p.length()) * normal;
+//	//dvec3 normal = unit_vector(rec.normal);
+//	dvec3 v2 = normal * dot(v, normal);
+//	dvec3 v1 = v - v2;
+//	dvec3 v1_p = (n_in / n_out) * v1.length() * unit_vector(v1);
+//	dvec3 v2_p = -sqrt(1-v1_p.length()*v1_p.length()) * normal;
 //	return v1_p + v2_p;
 //}
 
 /*
-double Lambertian::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, const vec3& vt) const
+double Lambertian::BxDF(const dvec3& vi, double wli, const dvec3& vo, double wlo, const dvec3& vt) const
 {
 	if (vi.z() < 0.0)
 		return 0.0;
@@ -128,11 +128,11 @@ double Lambertian::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, 
 	return albedo.get(wli)/M_PI;
 }
 
-bool Lambertian::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF, double& pdfval) const
+bool Lambertian::Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF, double& pdfval) const
 {
 	CosinePdf Pdf(rec.normal);
 
-	vec3 generated_direction = Pdf.Generate();
+	dvec3 generated_direction = Pdf.Generate();
 	pdfval = Pdf.PdfVal(generated_direction);
 	vi = uvw.WorldToLocal(generated_direction);
 
@@ -142,14 +142,14 @@ bool Lambertian::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 }
 
 
-double Metal::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, const vec3& vt) const
+double Metal::BxDF(const dvec3& vi, double wli, const dvec3& vo, double wlo, const dvec3& vt) const
 {
 	if (vi.z() < 0.0)
 		return 0.0;
 
 	return albedo.get(wli) / abs(vi.z());
 }
-bool Metal::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF, double& pdfval) const
+bool Metal::Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF, double& pdfval) const
 {
 	double ref_idx = n.get(wlo);
 	double kv = k.get(wlo);
@@ -195,7 +195,7 @@ double Metal::BxDF(const ray& r_in, const HitRecord& rec, const ray& scattered) 
 	if (cosine < 0)
 		return 0;
 
-	vec3 reflected = reflect(unit_vector(r_in.direction()), rec.normal);
+	dvec3 reflected = reflect(unit_vector(r_in.direction()), rec.normal);
 
 	return albedo.get(r_in.central_wl) * 3.0*pow(dot(unit_vector(scattered.direction()), unit_vector(reflected)), 100) / M_PI;
 	return pow(dot(unit_vector(scattered.direction()), unit_vector(reflected)), 100);
@@ -203,10 +203,10 @@ double Metal::BxDF(const ray& r_in, const HitRecord& rec, const ray& scattered) 
 
 bool Metal::Sample(const ray& r_in, const HitRecord& rec, ray& scattered, double& BxDF, double& Pdf) const
 {
-	vec3 normal = unit_vector(rec.normal);
-	vec3 v = unit_vector(r_in.direction());
+	dvec3 normal = unit_vector(rec.normal);
+	dvec3 v = unit_vector(r_in.direction());
 	double abs_cos_o = abs(dot(v, normal));
-	vec3 r_out = v + 2*abs_cos_o*normal;
+	dvec3 r_out = v + 2*abs_cos_o*normal;
 	scattered = ray(rec.p, unit_vector(r_out));
 	scattered.min_wl = r_in.min_wl;
 	scattered.max_wl = r_in.max_wl;
@@ -219,10 +219,10 @@ bool Metal::Sample(const ray& r_in, const HitRecord& rec, ray& scattered, double
 */
 
 /*
-bool Metal::scatter(const ray& r_in, const HitRecord& rec, vec3& attenuation, ray& scattered, double& Pdf) const
+bool Metal::scatter(const ray& r_in, const HitRecord& rec, dvec3& attenuation, ray& scattered, double& Pdf) const
 {
-	vec3 v = unit_vector(r_in.direction());
-	vec3 reflected = reflect(v, rec.normal);
+	dvec3 v = unit_vector(r_in.direction());
+	dvec3 reflected = reflect(v, rec.normal);
 	scattered = ray(rec.p, reflected + fuzz*random_in_unit_Sphere());
 	attenuation = albedo;
 	//return (dot(scattered.direction(), rec.normal) > 0);
@@ -240,7 +240,7 @@ bool Metal::scatter(const ray& r_in, const HitRecord& rec, vec3& attenuation, ra
 
 
 /*
-bool Dielectric::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF, double& pdfval) const
+bool Dielectric::Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF, double& pdfval) const
 {
 
 	//double ref_idx = ref_B + ref_C / pow(r_in.central_wl/1000.0, 2.0);
@@ -251,14 +251,14 @@ bool Dielectric::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 	double n_vacuum = 1.0;
 	std::complex<double> n1, n2;
 	double fresnel;
-	vec3 normal;
+	dvec3 normal;
 	if (cos_o >= 0.0) {
 		// from outside to inside of object
 		//n_second = ref_idx;
 		//n_first = n_vacuum;
 		n1 = n_vacuum;
 		n2 = std::complex<double>(ref_idx, kv);
-		normal = vec3(0, 0, 1);
+		normal = dvec3(0, 0, 1);
 	} else {
 		// from inside to outside of object
 		//n_second = n_vacuum;
@@ -266,7 +266,7 @@ bool Dielectric::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 		n1 = std::complex<double>(ref_idx, kv);
 		n2 = n_vacuum;
 		cos_o = abs(cos_o);
-		normal = vec3(0, 0, -1);
+		normal = dvec3(0, 0, -1);
 	}
 	//double relative_ref_idx = std::real(n2)/std::real(n1);
 	std::complex<double> relative_ref_idx = n2/n1;
@@ -309,7 +309,7 @@ bool Dielectric::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 		double sin_o = sqrt(std::max(0.0, 1.0-cos_o*cos_o));
 		double sin_t = sin_o / std::real(relative_ref_idx);
 		double cos_t = sqrt(1.0-sin_t*sin_t);
-		vi = (-cos_t) * normal - sin_t * unit_vector(vec3(vo[0], vo[1], 0));
+		vi = (-cos_t) * normal - sin_t * unit_vector(dvec3(vo[0], vo[1], 0));
 		//pdfval = 1.0-fresnel;
 		//BxDF = ((n_first*n_first)/(n_second*n_second)) * (1.0-fresnel) / cos_t;
 		pdfval = 1.0;
@@ -321,7 +321,7 @@ bool Dielectric::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 	return true;
 }
 
-double OrenNayar::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, const vec3& vt) const
+double OrenNayar::BxDF(const dvec3& vi, double wli, const dvec3& vo, double wlo, const dvec3& vt) const
 {
 	double cos_theta_i = vi.z();
 	if (cos_theta_i < 0) {
@@ -336,14 +336,14 @@ double OrenNayar::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, c
 	double A = 1 - sigma*sigma/(2*(sigma*sigma + 0.33));
 	double B = 0.45 * sigma * sigma / (sigma*sigma+0.09);
 
-	vec3 hori_vo = vec3(vo[0], vo[1], 0);
-	vec3 hori_scattered = vec3(vi[0], vi[1], 0);
+	dvec3 hori_vo = dvec3(vo[0], vo[1], 0);
+	dvec3 hori_scattered = dvec3(vi[0], vi[1], 0);
 	double tmp = std::max(0.0, dot(hori_vo, hori_scattered));
 
 	return albedo.get(wli)/M_PI * (A + B * tmp * sin_alpha * tan_beta);
 }
 
-bool OrenNayar::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF, double& pdfval) const
+bool OrenNayar::Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF, double& pdfval) const
 {
 	//std::vector<std::unique_ptr<Pdf> > Pdf_list(lights.size()+1);
 	//Pdf_list[0] = std::make_unique<UniformPdf>(rec.normal);
@@ -352,7 +352,7 @@ bool OrenNayar::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, dou
 	//}
 	//MixturePdf Pdf(std::move(Pdf_list));
 
-	//vec3 generated_direction = Pdf.Generate();
+	//dvec3 generated_direction = Pdf.Generate();
 	//pdfval = Pdf.PdfVal(generated_direction);
 	//vi = uvw.WorldToLocal(generated_direction);
 
@@ -363,14 +363,14 @@ bool OrenNayar::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, dou
 }
 
 
-double Microfacet::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, const vec3& vt) const
+double Microfacet::BxDF(const dvec3& vi, double wli, const dvec3& vo, double wlo, const dvec3& vt) const
 {
 	double cos_theta_i = vi.z();
 	if (cos_theta_i < 0) {
 		return 0.0;
 	}
 
-	vec3 vh = (vi+vo)/2.0;
+	dvec3 vh = (vi+vo)/2.0;
 
 	//double cos_theta_h_squared = vh.z() * vh.z();
 	//double tan_theta_h_squared = 1.0/cos_theta_h_squared - 1.0;
@@ -392,7 +392,7 @@ double Microfacet::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, 
 	return albedo.get(wli) * d * g * fresnel / (4.0 * vo.z() * vh.z());
 }
 
-double Microfacet::lambda(const vec3& v) const
+double Microfacet::lambda(const dvec3& v) const
 {
 	double cos_theta_squared = v.z()*v.z();
 	double tan_theta = sqrt(1.0/cos_theta_squared - 1.0);
@@ -401,7 +401,7 @@ double Microfacet::lambda(const vec3& v) const
 }
 
 
-bool Microfacet::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF, double& pdfval) const
+bool Microfacet::Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF, double& pdfval) const
 {
 	//double phi = 2.0 * M_PI * drand48();
 	//double tan_theta = sqrt(-alpha*alpha*log(1.0-drand48()));
@@ -417,7 +417,7 @@ bool Microfacet::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 	//double sin_theta = sqrt(1.0 - cos_theta_squared);
 	//
 
-	vec3 micro_normal;
+	dvec3 micro_normal;
 	micro_normal[0] = sin_theta * cos(phi);
 	micro_normal[1] = sin_theta * sin(phi);
 	micro_normal[2] = sqrt(cos_theta_squared);
@@ -478,14 +478,14 @@ bool Microfacet::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, do
 	return true;
 }
 
-double GGX::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, const vec3& vt) const
+double GGX::BxDF(const dvec3& vi, double wli, const dvec3& vo, double wlo, const dvec3& vt) const
 {
 	double cos_theta_i = vi.z();
 	if (cos_theta_i < 0) {
 		return 0.0;
 	}
 
-	vec3 vh = (vi+vo)/2.0;
+	dvec3 vh = (vi+vo)/2.0;
 
 	double cos_theta_h_squared = vh.z() * vh.z();
 	double tan_theta_h_squared = 1.0/cos_theta_h_squared - 1.0;
@@ -500,7 +500,7 @@ double GGX::BxDF(const vec3& vi, double wli, const vec3& vo, double wlo, const v
 	return albedo.get(wli) * d * g * fresnel / (4.0 * vo.z() * vi.z());
 }
 
-double GGX::lambda(const vec3& v) const
+double GGX::lambda(const dvec3& v) const
 {
 	double cos_theta_squared = v.z()*v.z();
 	double tan_theta = sqrt(1.0/cos_theta_squared - 1.0);
@@ -508,11 +508,11 @@ double GGX::lambda(const vec3& v) const
 	return 0.5 * (std::erf(a) - 1.0 + exp(a*a) / (a * sqrt(M_PI)));
 }
 
-bool GGX::Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF, double& pdfval) const
+bool GGX::Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF, double& pdfval) const
 {
 	CosinePdf Pdf(rec.normal);
 
-	vec3 generated_direction = Pdf.Generate();
+	dvec3 generated_direction = Pdf.Generate();
 	pdfval = Pdf.PdfVal(generated_direction);
 	vi = uvw.WorldToLocal(generated_direction);
