+++ material.h
@@ -9,8 +9,8 @@
 #include "hittable.h"
 #include "spectrum.h"
 
-extern const vec3 default_vt;
-vec3 reflect2(vec3 v, vec3 normal);
+extern const dvec3 default_vt;
+dvec3 reflect2(dvec3 v, dvec3 normal);
 double rfresnel(double cos_theta, double n);
 double cfresnel(double cos_theta, std::complex<double> n);
 
@@ -18,10 +18,10 @@ class MediumMaterial {
 	public:
 		MediumMaterial(const Spectrum& sigma_t, const Spectrum& albedo) : sigma_t(sigma_t), albedo(albedo) {
 		}
-		virtual bool Sample_p(const vec3& vo, double wlo, vec3& vi, double& wli, double& phase, double& pdfval) const {
+		virtual bool Sample_p(const dvec3& vo, double wlo, dvec3& vi, double& wli, double& phase, double& pdfval) const {
 			return false;
 		}
-		virtual double Phase(const vec3& vi, double wli, const vec3& vo, double wlo) const = 0;
+		virtual double Phase(const dvec3& vi, double wli, const dvec3& vo, double wlo) const = 0;
 		Spectrum sigma_t;
 		Spectrum albedo;
 };
@@ -29,15 +29,15 @@ class MediumMaterial {
 class Homogenious : public MediumMaterial {
 	public:
 		Homogenious(const Spectrum& sigma_t, const Spectrum& albedo) : MediumMaterial(sigma_t, albedo) {}
-		bool Sample_p(const vec3& vo, double wlo, vec3& vi, double& wli, double& phase, double& pdfval) const override;
-		double Phase(const vec3& vi, double wli, const vec3& vo, double wlo) const override;
+		bool Sample_p(const dvec3& vo, double wlo, dvec3& vi, double& wli, double& phase, double& pdfval) const override;
+		double Phase(const dvec3& vi, double wli, const dvec3& vo, double wlo) const override;
 };
 
 class HenyeyGreenstein : public MediumMaterial {
 	public:
 		HenyeyGreenstein(const Spectrum& sigma_t, const Spectrum& albedo, double g) : MediumMaterial(sigma_t, albedo), g(g) {}
-		bool Sample_p(const vec3& vo, double wlo, vec3& vi, double& wli, double& phase, double& pdfval) const override;
-		double Phase(const vec3& vi, double wli, const vec3& vo, double wlo) const override;
+		bool Sample_p(const dvec3& vo, double wlo, dvec3& vi, double& wli, double& phase, double& pdfval) const override;
+		double Phase(const dvec3& vi, double wli, const dvec3& vo, double wlo) const override;
 		double g;
 };
 
@@ -47,13 +47,13 @@ class Material {
 		virtual void PreProcess(HitRecord &rec) const
 		{
 		}
-		virtual bool Sample(const HitRecord& rec, const ONB& uvw, const vec3& vo, double wlo, vec3& vi, double& wli, double& BxDF_divided_by_pdf, double& BxDF, double& pdfval) const {
+		virtual bool Sample(const HitRecord& rec, const ONB& uvw, const dvec3& vo, double wlo, dvec3& vi, double& wli, double& BxDF_divided_by_pdf, double& BxDF, double& pdfval) const {
 			return false;
 		}
-		virtual double BxDF(const vec3& vi, double wli, const vec3& vo, double wlo) const {
+		virtual double BxDF(const dvec3& vi, double wli, const dvec3& vo, double wlo) const {
 			return 0.0;
 		}
-		virtual double PDF(const vec3& vi, double wli, const vec3& vo, double wlo) const {
+		virtual double PDF(const dvec3& vi, double wli, const dvec3& vo, double wlo) const {
 			return 0.0;
 		}
 		virtual double Emitted(const ray& r, const HitRecord& rec) const {
